Update Steam patch; Refactor; Conditionally keep window open
Swapped the Steam patch for the new update as well as changed the method
The spot is the same, but instead of changing jnz -> jz, we use jnz ->
nop jmp.

Refactored the memory functions so that the patch, memory, and offset
are decoupled from themselves in order to make them more generic.

If the patcher is run from outside a CLI, it will now stay open until
a character is pressed, allowing error messages and output to be seen
more easily.

Lastly, the patcher now checks to make sure that the patch was applied
correctly, by applying the patch to a duplicate of the found array and
seeing that it matches what's in memory.

Closes #1.