Swap to dynamically finding patch
Instead of finding a given set of assembly bytes, now we find the string
which appears when downloading a depot fails, calculate the offset used
for that string, and then find the assembly `pop offset`. We then go
backwards and find the `jnz addr` before the pop which branches away
from the pop of the string on failure and change it to a `nop, jmp`.